Remove RmrIntegrationTest. Move tests into RmIntegrationTest
