+++ tests/src/test/java/alluxio/client/cli/fs/command/RmCommandIntegrationTest.java
@@ -80,4 +80,53 @@ public final class RmCommandIntegrationTest extends AbstractFileSystemShellTest 
     Assert.assertTrue(fileExists(new AlluxioURI(testDir + "/bar")));
     Assert.assertTrue(fileExists(new AlluxioURI(testDir + "/bar/foobar3")));
   }
+
+  @Test
+  public void rmr() throws IOException {
+    StringBuilder toCompare = new StringBuilder();
+    mFsShell.run("mkdir", "/testFolder1/testFolder2");
+    toCompare.append(getCommandOutput(new String[] {"mkdir", "/testFolder1/testFolder2"}));
+    mFsShell.run("touch", "/testFolder1/testFolder2/testFile2");
+    toCompare
+        .append(getCommandOutput(new String[] {"touch", "/testFolder1/testFolder2/testFile2"}));
+    AlluxioURI testFolder1 = new AlluxioURI("/testFolder1");
+    AlluxioURI testFolder2 = new AlluxioURI("/testFolder1/testFolder2");
+    AlluxioURI testFile2 = new AlluxioURI("/testFolder1/testFolder2/testFile2");
+    Assert.assertTrue(fileExists(testFolder1));
+    Assert.assertTrue(fileExists(testFolder2));
+    Assert.assertTrue(fileExists(testFile2));
+    mFsShell.run("rm", "-R", "/testFolder1/testFolder2/testFile2");
+    toCompare.append(getCommandOutput(new String[] {"rm", "/testFolder1/testFolder2/testFile2"}));
+    Assert.assertEquals(toCompare.toString(), mOutput.toString());
+    Assert.assertTrue(fileExists(testFolder1));
+    Assert.assertTrue(fileExists(testFolder2));
+    Assert.assertFalse(fileExists(testFile2));
+    mFsShell.run("rm", "-R", "/testFolder1");
+    toCompare.append(getCommandOutput(new String[] {"rmr", "/testFolder1"}));
+    Assert.assertEquals(toCompare.toString(), mOutput.toString());
+    Assert.assertFalse(fileExists(testFolder1));
+    Assert.assertFalse(fileExists(testFolder2));
+    Assert.assertFalse(fileExists(testFile2));
+  }
+
+  @Test
+  public void rmrWildCard() throws IOException, AlluxioException {
+    String testDir = FileSystemShellUtilsTest.resetFileHierarchy(mFileSystem);
+
+    mFsShell.run("rm",  "-R", testDir + "/foo/foo*");
+    Assert.assertFalse(fileExists(new AlluxioURI(testDir + "/foo/foobar1")));
+    Assert.assertFalse(fileExists(new AlluxioURI(testDir + "/foo/foobar2")));
+    Assert.assertTrue(fileExists(new AlluxioURI(testDir + "/foo")));
+    Assert.assertTrue(fileExists(new AlluxioURI(testDir + "/bar/foobar3")));
+
+    mFsShell.run("rm", "-R", testDir + "/ba*");
+    Assert.assertFalse(fileExists(new AlluxioURI(testDir + "/bar")));
+    Assert.assertFalse(fileExists(new AlluxioURI(testDir + "/bar/foobar3")));
+    Assert.assertTrue(fileExists(new AlluxioURI(testDir + "/foobar4")));
+
+    mFsShell.run("rm", "-R", testDir + "/*");
+    Assert.assertFalse(fileExists(new AlluxioURI(testDir + "/bar")));
+    Assert.assertFalse(fileExists(new AlluxioURI(testDir + "/foo")));
+    Assert.assertFalse(fileExists(new AlluxioURI(testDir + "/foobar4")));
+  }
 }
